requirements.txt
altgraph==0.17.2
cachelib==0.4.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Session==0.4.0
future==0.18.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pefile==2021.9.3
protobuf==3.19.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pywin32-ctypes==0.2.0
Werkzeug==2.0.2
